requirements.txt
python-dotenv==1.1.0
aiogram==3.20.0.post0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
yt-dlp==2025.8.27